Filter builder updates (#11)
- Updated some tests to cover the changes.
- Updated the date time format to be 24-hour format as that's the format
jotform filter string expects.
- Fixed the date filters logic, currently jotform filter commands don't
support GreaterThanOrEqualTo (>=) and LessThanOrEqualTo (<=), which
means the date range filter requires adjusting the dates. To make the
date range filter convenient I last time added optional flags for
adding/subtracting one day from dates but that only worked if the time
is zero in the dates, so this update is an attempt to address that
limitation by subtracting only 1 second and adding either 1 second or
day depends if the time is included in the date.